整数計算による計算誤差から起きていたバグを修正

+++ Inc/target_based_on_time/target_based_on_time.hpp
@@ -25,8 +25,8 @@ private:
 
 	float position;
 	float velocity;
-	float getPositionBasic(uint16_t time);
-	float getVelocityBasic(uint16_t time);
+	float getPositionBasic(float time);
+	float getVelocityBasic(float time);
 
 	float periodOfAcceleration; //加速時間
 	float periodOfConstantVelocity; //等速時間
